[refactor] Resolve Conflict
