+++ src/main/java/gp/cnusambe/dto/LectureMapDto.java
@@ -1,21 +1,18 @@
-package gp.cnusambe.payload.response;
+package gp.cnusambe.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import gp.cnusambe.dto.LectureSWDto;
-import gp.cnusambe.dto.SWInLectureSWDto;
+import gp.cnusambe.domain.LectureMap;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import java.util.List;
 
 @Setter
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class LectureSWResponse {
-    private LectureSWDto lectureSW;
-    private List<SWInLectureSWDto> sw;
+public class LectureMapDto {
+    private LectureMap lectureMap;
 }
