+++ src/main/java/gp/cnusambe/dto/SWInLectureSWDto.java
@@ -2,6 +2,7 @@ package gp.cnusambe.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
@@ -9,8 +10,10 @@ import lombok.Setter;
 @Setter
 @Getter
 @NoArgsConstructor
+@AllArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class SWInLectureSWDto {
+    private Long registrationSWId;
     private String swManufacturer;
     private String swName;
 }
